+++ src/main/java/org/example/Expr.java
@@ -3,5 +3,5 @@ package org.example;
 import java.util.Map;
 
 public interface Expr extends Node{
-    double eval(Map<String, Double> bindings) throws EvalError;
+    double eval(Map<String, Integer> bindings) throws EvalError;
 }
